requirements.txt
# CFST XGBoost Pipeline Requirements

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
xgboost>=2.0.0
scikit-learn>=1.3.0

# Hyperparameter optimization
optuna>=3.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.12.0

# Configuration
pyyaml>=6.0

# Model serialization
joblib>=1.3.0

# Logging (Python standard library - no extra package needed)
